Refactor All DI modules inside a Modules.kt file (#299)
* Updated Fragment_media_details

Added Padding in category item and  file usage item

* SplashScreen test

* Updated

* Refactor All DI modules inside a Modules.kt file

* Converted PresenterCallback from java to Kotlin

* changed

* Update SplashScreenTest.java
